Javaslang 2.1.0 contains a breaking change of type from Long to Integer for collection indices. This requires the same change in type signature of Every.zipWithIndex().

+++ src/main/java/com/github/javactic/Every.java
@@ -1972,7 +1972,7 @@ public interface Every<T> extends Iterable<T>, IntFunction<T> {
    * @return A new Every containing pairs consisting of all elements of this
    *         Every paired with their index. Indices start at 0.
    */
-  default Every<Tuple2<T, Long>> zipWithIndex() {
+  default Every<Tuple2<T, Integer>> zipWithIndex() {
     return fromNonEmptySeq(toVector().zipWithIndex());
   }
